feat(signup): complete signup page v2.5
