+++ app/src/main/java/com/example/newgroupapplication/views/auth/SignUp/SignUpFragment.kt
@@ -1,0 +1,19 @@
+package com.example.newgroupapplication.views.auth.SignUp
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.newgroupapplication.R
+
+class SignUpFragment : Fragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_sign_up, container, false)
+    }
+
+}
